added emergency stop

+++ mrobosub_hal/src/thruster_controller.py
@@ -7,6 +7,7 @@ from std_msgs.msg import Float32
 from serial import Serial
 from typing_extensions import Callable
 from mrobosub_msgs.msg import MotorState
+from std_srvs.srv import SetBool, SetBoolResponse
 
 NUM_MOTORS = 8
 
@@ -18,20 +19,37 @@ class ThrusterController(Node):
         # self.port = '/dev/ttyACM0'
         self.get_errors() # clear errors at the start
         
+        self.object_position_service = rospy.Service("emergency_stop_motors", SetBool, self.handle_emergency_stop)
+
         self.motor_sub = rospy.Subscriber('/motor_output', MotorState, self.motor_callback)
+        self.emergency_stop = False
         rospy.spin()
     
+    def handle_emergency_stop(self, _):
+        self.emergency_stop = True
+        self.get_errors()
+        try:
+            with Serial(self.port) as s:
+                s.write(bytearray([0xAA, 0x0C, 0x22]))
+        except:
+            self.get_errors() #giving it one more chance to clear errors, just in case (sometimes it's weird)
+            with Serial(self.port) as s:
+                s.write(bytearray([0xAA, 0x0C, 0x22]))
+        r = SetBoolResponse()
+        r.success = True
+        return r
+
     # pwm_raw ranges from -1 to 1
     # pwm_val ranges from 4000 to 8000
-    def convert_pwm_signal(self, pwm_raw: Float32) -> int:
-        if(pwm_raw.data<-1 or pwm_raw.data>1):
+    def convert_pwm_signal(self, pwm_raw: float) -> int:
+        if(pwm_raw<-1 or pwm_raw>1):
             print(f"Thruster Controller [ERROR]: PWM value {pwm_raw} out of range (should be in [-1, 1])")
             return -1
-        return int((pwm_raw.data * 2000)+6000)
+        return int((pwm_raw * 2000)+6000)
     
     # in case of invalid PWM or motor number parameters, does not send any updated signal to the motor controller and returns -1
     def send_signal(self, motor: int, pwm_raw: Float32) -> int:
-        pwm_val: int = self.convert_pwm_signal(pwm_raw)
+        pwm_val: int = self.convert_pwm_signal(pwm_raw.data)
         if pwm_val == -1:
             return -1
         
@@ -56,9 +74,10 @@ class ThrusterController(Node):
         return 0
 
     def motor_callback(self, msg: MotorState):
-        for i in range(NUM_MOTORS):
-            motor_name = f"motor{i}"
-            self.send_signal(i, msg.motor_name)
+        if not self.emergency_stop:
+            for i in range(NUM_MOTORS):
+                motor_name = f"motor{i}"
+                self.send_signal(i, msg.motor_name)
 
     def get_errors(self):
         # gets errors from thruster controller hardware (which automatically clears the errors too)
@@ -69,6 +88,7 @@ class ThrusterController(Node):
             if(error_code != 0):
                 print(f"Thruster controller: error code = {error_code}")
             # eg: error_code 16 means 00010000 which is the 5th error bit set
+
 
 
 def main():
